+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## New data with motivation
 
+* goulden.barley.uniformity (expanded from 20x20 to 48x48)
 * smith.wheat.uniformity (classic data for measuring field heterogeneity)